chore: fix cyclic dependency in cmap module
