+++ test/unit/cmap/connection_pool.test.js
@@ -6,7 +6,7 @@ const { ConnectionPool } = require('../../../src/cmap/connection_pool');
 const { WaitQueueTimeoutError } = require('../../../src/cmap/errors');
 const { EventEmitter } = require('events');
 const mock = require('../../tools/mock');
-const cmapEvents = require('../../../src/cmap/events');
+const cmapEvents = require('../../../src/cmap/connection_pool_events');
 const sinon = require('sinon');
 const { expect } = require('chai');
 const { ns } = require('../../../src/utils');
